move message sending files to a separate package
