+++ crash_alert/message_senders/message_sender_base.py
